+++ Semana_3/githubrepo/src/App.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import Routes from "./routes";
+import { BrowserRouter } from "react-router-dom";
+import GlobalStyle from "./Styles/global";
+
+const App: React.FC = () => (
+  <>
+    <BrowserRouter>
+      <Routes />
+    </BrowserRouter>
+    <GlobalStyle/>
+  </>
+);
+export default App;