+++ Semana_3/githubrepo/src/index.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import App from './App';
+
+
+ReactDOM.render(
+  <React.StrictMode>     
+    <App />
+  </React.StrictMode>,
+  document.getElementById('root')
+);
+
+//React.StrictMode - Habilita uma tratativa de erros diferente, otimizada para TS
+
+
+