+++ Semana_3/githubrepo/src/react-app-env.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="react-scripts" />
+
+// Arquivo para o vs code entender as tipagens do react